Make a few provisioning commands idempotent for robustness
Also add a bunch of nodes regarding provisioning.  Was hoping to fully
automate it with this patch but there are still a few things that won't
automate and this thing went from "Real quick" to 4 hours, so I'm
punting on that for later.  At least with this it can be done entirely
with the CLI and don't have to open the aws console

+++ scripts/provision-build-arm-release.sh
@@ -1,9 +1,55 @@
 #!/usr/bin/env bash
 
 #
-# This script is meant to run on s fresh ubuntu ARM64 ec2 instance.
+# This script is meant to run on a fresh ubuntu ARM64 ec2 instance.
 # It will build and push malan latest arm releases
 #
+# To find the image ID (not this takes a LONG time it's 262MB!):
+#   aws ec2 describe-images | jq -r ".Images" > images.json
+#
+# To use irb:
+#   irb
+#   require 'json'
+#   require 'date'
+#   puts JSON.parse(File.read("images.json"))
+#      .filter{|i| i["OwnerId"] == "099720109477" && i["Architecture"] == "arm64" && i["State"] == "available" && i["Name"] =~ /ubuntu-jammy/i && i["Name"] =~ /^ubuntu\/images\//i}
+#      .sort{|i1, i2| DateTime.parse(i2["CreationDate"]) <=> DateTime.parse(i1["CreationDate"])}
+#      .map{|i| "#{i["ImageId"]} - #{i["Name"]}"}
+#      .join("\n")
+#
+# To use ruby from the CLI:
+#   ruby -r 'json' -r 'date' -e 'puts JSON.parse(File.read("images.json")).filter{|i| i["OwnerId"] == "099720109477" && i["Architecture"] == "arm64" && i["State"] == "available" && i["Name"] =~ /ubuntu-jammy/i && i["Name"] =~ /^ubuntu\/images\//i}.sort{|i1, i2| DateTime.parse(i2["CreationDate"]) <=> DateTime.parse(i1["CreationDate"])}.map{|i| "#{i["ImageId"]} - #{i["Name"]}"}.join("\n")'
+
+# To get the image ID:
+#
+#   ruby -r 'json' -r 'date' -e 'puts JSON.parse(File.read("images.json")).filter{|i| i["OwnerId"] == "099720109477" && i["Architecture"] == "arm64" && i["State"] == "available" && i["Name"] =~ /ubuntu-jammy/i && i["Name"] =~ /^ubuntu\/images\//i}.sort{|i1, i2| DateTime.parse(i2["CreationDate"]) <=> DateTime.parse(i1["CreationDate"])}.first.to_json' | jq -r '.ImageId'
+#
+#
+# # aws ec2 run-instances --image-id ami-0ee02425a4c7e78bb --count 1 --instance-type c6g.xlarge --key-name ben_0 --security-group-ids sg-903004f8 --subnet-id subnet-6e7f829e
+#
+# imageid="$(ruby -r 'json' -r 'date' -e 'puts JSON.parse(File.read("images.json")).filter{|i| i["OwnerId"] == "099720109477" && i["Architecture"] == "arm64" && i["State"] == "available" && i["Name"] =~ /ubuntu-jammy/i && i["Name"] =~ /^ubuntu\/images\//i}.sort{|i1, i2| DateTime.parse(i2["CreationDate"]) <=> DateTime.parse(i1["CreationDate"])}.first.to_json' | jq -r '.ImageId')"
+#
+# aws ec2 run-instances --image-id ami-0ee02425a4c7e78bb --count 1 --instance-type c6g.xlarge --key-name ben_0
+# # wait for instance to initialize and get a public IP
+# # aws ec2 describe-instances | jq -r '.Reservations | map(.Instances) | map(.PublicIpAddress)'
+# # Get the public IP of the machine
+# iid="$(aws ec2 describe-instances | jq -r '.Reservations[].Instances[].InstanceId')"
+# vmip="$(aws ec2 describe-instances | jq -r '.Reservations[].Instances[].PublicIpAddress')"
+# sgid="$(aws ec2 describe-instances | jq -r '.Reservations[].Instances[].SecurityGroups[].GroupId')"
+# # Allow SSH in the security group
+# aws ec2 authorize-security-group-ingress --group-id "${sgid}" --protocol tcp --port 22 --cidr 0.0.0.0/0
+# # Copy docker creds
+# ssh ubuntu@${vmip} 'mkdir -p /home/ubuntu/.docker && sudo mkdir -p /root/.docker'
+# scp /home/ben/.docker/config.json ubuntu@${vmip}:/home/ubuntu/.docker/config.json
+# ssh ubuntu@${vmip} 'sudo cp /home/ubuntu/.docker/config.json /root/.docker/config.json'
+# # Invoke the script
+# scp scripts/provision-build-arm-release.sh ubuntu@${vmip}:/home/ubuntu/
+# # automatic doesn't work due to a terminal issue
+# # ssh ubuntu@${vmip} 'sudo ./provision-build-arm-release.sh'
+# ssh ubuntu@${vmip}
+#   sudo ./provision-build-arm-release.sh
+# # Destroy the instance
+# aws ec2 terminate-instances --instance-ids "${iid}"
 
 set -e
 
@@ -53,10 +99,14 @@ sudo apt-get update
 sudo apt-get -y install docker-ce docker-ce-cli containerd.io docker-compose-plugin
 
 # Setup docker auth
-mkdir -p $HOME/.docker
-echo "${DOCKER_CONFIG}" > $HOME/.docker/config.json
+if ! [ -f "$HOME/.docker/config.json" ]; then
+  mkdir -p $HOME/.docker
+  echo "${DOCKER_CONFIG}" > $HOME/.docker/config.json
+fi
 
-git clone https://github.com/FreedomBen/malan.git
+if ! [ -d "malan" ]; then
+  git clone https://github.com/FreedomBen/malan.git
+fi
 
 cd malan
